[UPDATE]优化仿真程序代码结构：拆分投注策略模块
