+++ NextBFootBallAnalysis/libs/sqlite_db.py
@@ -209,6 +209,26 @@ class NextbFootballSqliteDB:
         else:
             return []
 
+    def get_season_matchs(self, season):
+        """
+        获取指定赛季的全部比赛的结果
+        """
+        data = (
+            self.session_maker.query(NextbFootballDatas)
+            .filter(
+                NextbFootballDatas.season == season,
+            )
+            .order_by(NextbFootballDatas.date_time.asc())
+            .all()
+        )
+        if len(data) > 0:
+            datas = list()
+            for d in data:
+                datas.append(d)
+            return datas
+        else:
+            return []
+
     def get_team_last_season_matchs(self, team, number=5):
         """
         获取指定球队最近number个赛季的比赛结果，默认最近5个赛季
