+++ NextBFootBallAnalysis/libs/strategy/__init__.py
@@ -1,9 +1,7 @@
 # -*- coding: utf-8 -*-
-# @Time     : 2023/01/03 16:08:40
+# @Time     : 2023/02/22 17:53:04
 # @Author   : ddvv
 # @Site     : https://ddvvmmzz.github.io
 # @File     : __init__.py
 # @Software : Visual Studio Code
 # @WeChat   : NextB
-
-NEXTB_FOOTBALL_VERSION = "2.0.5"
